refractor： refactor role visivlble
